+++ SoftGL/GLObjects/ShaderProgram/FragmentShader.cs
@@ -3,7 +3,6 @@ using System;
 using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 
 namespace SoftGL
@@ -12,19 +11,10 @@ namespace SoftGL
     {
         public FragmentShader(uint id) : base(ShaderType.FragmentShader, id) { }
 
-        protected override string AfterCompile(Assembly assembly)
+
+        protected override string DoCompile()
         {
-            Type codeType = this.FindShaderCodeType(assembly, typeof(FragmentShaderCode));
-            if (codeType == null) { return "No FragmentShader found!"; }
-
-            {
-                Dictionary<string, UniformVariable> dict;
-                string result = FindUniformVariables(codeType, out dict);
-                if (result != string.Empty) { return result; }
-                this.uniformVariableDict = dict;
-            }
-
-            return string.Empty;
+            throw new NotImplementedException();
         }
     }
 }
