+++ SoftGL/GLObjects/ShaderProgram/ShaderProgram.Link.cs
@@ -23,13 +23,6 @@ namespace SoftGL
             throw new NotImplementedException();
         }
 
-        private bool FindUniforms()
-        {
-            bool result = true;
-
-            return result;
-        }
-
         /// <summary>
         /// find the vertex shader and other shaders.
         /// </summary>
